Refactor encryped secrets
